+++ dpm/distributions/dirichlet.py
@@ -15,7 +15,7 @@ class Dirichlet(Distribution):
         if not isinstance(alpha, torch.Tensor):
             alpha = torch.tensor(alpha).view(-1)
         self.n_dims = len(alpha)
-        self._alpha = utils.softplus_inverse(alpha)
+        self._alpha = utils.softplus_inverse(alpha.float())
         if learnable:
             self._alpha = Parameter(self._alpha)
